pdx206: Adjust BOARD_SUPER_PARTITION_SIZE
The Japanese variant SOG01 has a different super partition size, which
is smaller than the international variant (0x2C0000000 instead of 0x2E0000000).

Change the BOARD_SUPER_PARTITION_SIZE to match Japanese variant so that
it can flash successfully. This change does not require a data wipe so
it is safe.